requirements.txt
pandas==1.3.5
numpy==1.22.0
seaborn==0.11.2
matplotlib==3.5.1
scikit-learn==1.0.2
pyyaml==6.0
